requirements.txt
Flask_Bootstrap==3.3.7.1
Flask==1.0
WebUI==0.2.0
waitress==1.4.2
PyYAML==5.3
